1Refactor Redux actions and components for clarity and optimization
This commit does three things: 1) it converts the Redux action 'SET_INITIALIZES' to 'SET_INITIALIZED' for consistency and puts it inside an ActionTypes object for easier management; 2) it changes the 'getInitializeApp' function to a simpler, more concise 'initializeApp' function which does away with the Promise and directly dispatches the necessary actions; and 3) it removes unnecessary componentDidMount call in HeaderContainer, streamlining the app initialization process.

These changes enhance the readability and optimize the performance of the code.