Refactor pagination in User component
Added a dedicated Pagination component to handle the logic for dividing user list into pages, removing the responsibility from the User component. This modification is intended to decouple responsibilities, thereby enabling the Pagination component to be reused in other parts of the application if necessary. Additionally, logic for paginated views has been simplified and moved to the new Pagination component which has been stylized according to the application's design guidelines. The users.module.scss file was adjusted accordingly. WatcherTasks file was also added to enhance project configuration for development.