1"Refactor ProfileContainer, add photo upload feature"
The ProfileContainer has been refactored from a Class component into a Function component, enabling the use of hooks for better code management. A feature for photo upload was added where if the user doesn't have a photo, they are given the option to upload a photo. This required creating a new Redux action and reducer (savePhoto and SET_PHOTO_SUCCESS, respectively) as well as a new API call (savePhoto). Additionally, CSS for input type="file" has been added in ProfileInfo.module.scss to improve its look.