Refactor ProfileDataForm and add TextFormField component
The ProfileDataForm underwent significant changes including adding name validation logic and using the Controller component from react-hook-form for greater flexibility and performance.

A TextFormField component was also created for a more reusable and semantic approach to validating and managing the state of TextField components, improving code manageability and ease of future updates.

These improvements aim to ensure better user data handling and enhance the readability of the application's codebase.