Extend ProfileDataForm component to enable saving and editing profile data
ProfileDataForm component has been extended to enable saving and editing profile data. This includes `fullName`, `lookingForAJob`, `lookingForAJobDescription`, `aboutMe` and all the contact fields. The changes involved adding new props - `saveProfileData` and `setEditMode`, and updating the form‘s onSubmit function to call these props on form submission. The TextField in the form for name has been updated to use ‘fullName’ instead of ‘name’. Additionally, handling of profile contacts has been added to render text fields for each contact field. These changes improve the user interface and user experience in the Profile component by enabling a user to see his current profile status and to adjust it accordingly.

Other changes include minor name changes in the action creators and reducers, and passing down newly created action creators to the ProfileDataForm component as props, to enable saving the profile data. These changes were necessary to ensure the appropriate actions are dispatched as a result of user interactions with the ProfileDataForm.

This update is crucial for the usability of the user profile feature, enabling users to have a more interactive experience with their profile data.