Update DialogsContainer with Authentication check
The DialogsContainer component has been updated significantly - the code has been refactored with organized import sections and comments in the native language. The component now uses the "connect" function from the "react-redux" library to connect to the Redux store and is wrapped with the "withAuthRedirect" HOC, improving user authentication by redirecting if the user is not authenticated. Lastly, the actions object from "dialogs-reducer.ts" is spread when passed to connect to make them accessible as props in the Dialogs component.