Add new webResources.xml file and update package-lock.json dependencies
Added a new file, webResources.xml, to the .idea folder to specify the web resources path for the IDE. This helps the IDE understand the structure of the project, improving code navigation and completion.

Updated several dependencies in package-lock.json to keep the software up to date, including adding @react-buddy/palette-antd and updating versions of @types/node, @types/react, @babel/plugin-proposal-private-property-in-object, and other dependencies. This ensures that the project is using the latest library versions, reducing the risk of bugs and security vulnerabilities.